FIX: Fjerner avsnittskiller fra beregning og utvider istedenfor den fra shared-components
