+++ packages/prosess-beregningsgrunnlag/src/components/beregningForm/BeregningForm.tsx
@@ -6,7 +6,7 @@ import { InjectedFormProps } from 'redux-form';
 import { createSelector } from 'reselect';
 import { isAksjonspunktOpen } from '@fpsak-frontend/kodeverk/src/aksjonspunktStatus';
 import aksjonspunktCodes from '@fpsak-frontend/kodeverk/src/aksjonspunktCodes';
-import { AksjonspunktHelpTextHTML, VerticalSpacer } from '@fpsak-frontend/shared-components';
+import { AksjonspunktHelpTextHTML, VerticalSpacer, AvsnittSkiller } from '@fpsak-frontend/shared-components';
 import { Column, Row } from 'nav-frontend-grid';
 import { behandlingForm } from '@fpsak-frontend/form';
 import faktaOmBeregningTilfelle from '@fpsak-frontend/kodeverk/src/faktaOmBeregningTilfelle';
@@ -33,7 +33,6 @@ import BeregningsresultatTable from '../beregningsresultatPanel/Beregningsresult
 
 import AksjonspunktBehandlerAT from '../arbeidstaker/AksjonspunktBehandlerAT';
 import AksjonspunktBehandlerFL from '../frilanser/AksjonspunktBehandlerFL';
-import AvsnittSkiller from '../redesign/AvsnittSkiller';
 import beregningStyles from '../beregningsgrunnlagPanel/beregningsgrunnlag.less';
 
 // ------------------------------------------------------------------------------------------ //
