+++ packages/shared-components/src/AvsnittSkiller.spec.tsx
@@ -6,7 +6,6 @@ import AvsnittSkiller from './AvsnittSkiller';
 describe('AvsnittSkiller', () => {
   it('skal vise skillelinje', () => {
     const wrapper = shallow(<AvsnittSkiller />);
-
     const row = wrapper.find('Row');
     expect(row).toHaveLength(1);
     const div = wrapper.find('div');
